Updated code
I don't remember what I did.

I updated indices to start at 0 instead of -1.

I also need to implement a way to reference a register's value.

+++ Compiler/src/include/types.h
@@ -52,8 +52,8 @@ typedef enum {
 
 typedef struct {
 	OPERATION_T operation;
-	uint32_t value1;
-	uint32_t value2;
+	uint16_t value1;
+	uint16_t value2;
 } TOKEN_T;
 
 struct LIST {
